pyproject.toml
[tool.poetry]
name = "drf-project"
version = "0.1.0"
description = ""
authors = ["<Anton Vedischev> <<vedishchevanton@gmail.com>>"]
readme = "README.md"
packages = [{include = "drf_project"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.3"
psycopg2-binary = "^2.9.6"
pillow = "^10.0.0"
djangorestframework = "^3.14.0"
django-filter = "^23.2"
djangorestframework-simplejwt = "^5.2.2"
coverage = "^7.2.7"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.2.0"
stripe = "^5.5.0"
django-crontab = "^0.7.1"
celery = "^5.3.1"
redis = "^4.6.0"
django-celery-beat = "^2.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
